// src/models/stationMessage.ts
export interface messageItemType {
  content: string;
  dustbin: boolean;
  haveRead: boolean;
  messageId: string;
  messageType: string;
  readTime: string;
  receivers: string[];
  receiversName: string[];
  recordId: string;
  send: boolean;
  sendTime: string;
  sender: string;
  senderName: string;
  title: string;
}